Reworked nav commands to only work within the app
