+++ src/Exceptionless.Web/ClientApp/src/routes/+layout.svelte
@@ -6,14 +6,11 @@
 
     import { page } from '$app/stores';
     import { setDefaultBaseUrl, setAccessTokenStore } from '$api/FetchClient';
-    import { accessToken, isAuthenticated } from '$api/auth';
+    import { accessToken } from '$api/auth';
     import { Toaster } from '$comp/ui/sonner';
-    import NavigationCommand from './(components)/NavigationCommand.svelte';
 
     import '../app.css';
-    import { routes, type NavigationItemContext, type NavigationItem } from './routes';
-    import { getMeQuery } from '$api/queries/users';
-    import { derived } from 'svelte/store';
+    import { routes } from './routes';
     import { isCommandOpen } from '$lib/stores/app';
 
     onMount(() => {
@@ -42,20 +39,12 @@
     });
 
     const queryClient = new QueryClient();
-    const userQuery = getMeQuery(queryClient);
-
-    // TODO: Pass these down into the sidebar.
-    const filteredRoutes = derived(userQuery, ($userResponse) => {
-        const context: NavigationItemContext = { authenticated: $isAuthenticated, user: $userResponse.data };
-        return routes.filter((route) => (route.show ? route.show(context) : true));
-    });
 </script>
 
 <div class="bg-background text-foreground">
     <ModeWatcher defaultMode={'dark'} />
 
     <QueryClientProvider client={queryClient}>
-        <NavigationCommand bind:open={$isCommandOpen} routes={$filteredRoutes} />
         <slot />
 
         <SvelteQueryDevtools />
